refactor: move image build script to cli
