+++ src/cli/utils/image-builder.js
